move overlay editor style file into cells folder
